data_storage/code/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
whitenoise==6.6.0
daphne==4.1.2
paho-mqtt==2.1.0
python-dateutil==2.8.2
channels==4.2.0
requests==2.32.3
pyzmq==26.2.0
django-admin-sortable==2.3
djangorestframework-csv==3.0.2